requirements.txt
flask==1.1.1
slackclient==2.1.0
gunicorn==19.9.0
psycopg2-binary==2.8.3
